Implement quickselect-based median algorithm
This calculates the median of a list of doubles in expected O(n) time,
but can experience worst-case behaviour if, for example, the list is
already sorted. It would be a good idea to use a randomly-chosen pivot
instead of always pivoting on the first element.

The idea is to use this for a faster pruning stage in HalfLifeMap -
rather than sorting the list of counts (O(n log n)), find the median and
then delete anything lower than it (O(n)).