QuickSelect-based HalfLifeMap (#2)
* Make HalfLifeMap tests more realistic

We're using maps with an idealSize of 25000 in Zonefox; test for that
case. Also increase the size of the keyspace to 26^4 - we don't know
exactly how large our keyspace is, but it's closer to 26^4 (4.5e6) than
26^2 (676).

* Implement quickselect-based median algorithm

This calculates the median of a list of doubles in expected O(n) time,
but can experience worst-case behaviour if, for example, the list is
already sorted. It would be a good idea to use a randomly-chosen pivot
instead of always pivoting on the first element.

The idea is to use this for a faster pruning stage in HalfLifeMap -
rather than sorting the list of counts (O(n log n)), find the median and
then delete anything lower than it (O(n)).

* Use select() in HalfLifeMap

* Delete empty Streaming(Mean|Variance)Test files

They contained commented-out test code for
StreamingMean/StreamingVariance, which appear never to have been
included in this repo. The presence of the empty test files was causing
errors when I try to build winterwell.maths.jar in Eclipse.