FrequencyCondDistribution: split out generic-map constructor
I want to pass a Context object to the normal map constructor so I'm not at the mercy of Cntxt.toString (which produces results that confuse my Redis GUI client). But if I make newMap take a Context as its argument then I get ClassCastExceptions when I construct the generic map, which passes in "!generic" and relies on erasure.